[CLEANUP] Split RestMeta LoadSaveTest
